@@ -1,18 +1,18 @@
-class ProductRoot < ActiveRecord::Base
+class User < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
-  include Patcher
-
   # security (i.e. attr_accessible) ...........................................
-  # attr_accessible :url
-
   # relationships .............................................................
   # validations ...............................................................
-  validates :url, :uniqueness => true
+  validates :username, uniqueness: true
 
   # callbacks .................................................................
   # scopes ....................................................................
   # additional config .........................................................
+  # Include default devise modules. Others available are:
+  # :confirmable, :lockable, :timeoutable and :omniauthable
+  devise :database_authenticatable, :registerable, :recoverable, :rememberable, :trackable, :validatable
+
   # class methods .............................................................
   # public instance methods ...................................................
   # protected instance methods ................................................
