@@ -2,7 +2,7 @@ class Bargain < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :price, :discount, :history_low
+  # attr_accessible :price, :discount, :history_low
 
   # relationships .............................................................
   belongs_to :product
