@@ -2,7 +2,7 @@ class BargainsCategory < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :bargain_id, :category_id
+  # attr_accessible :bargain_id, :category_id
 
   # relationships .............................................................
   belongs_to :bargain
