@@ -2,7 +2,7 @@ class Category < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :name, :words
+  # attr_accessible :name, :words
 
   # relationships .............................................................
   # validations ...............................................................
