@@ -2,11 +2,11 @@ class Price < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :value
+  # attr_accessible :value
 
   # relationships .............................................................
   belongs_to :product
-  
+
   # validations ...............................................................
   # callbacks .................................................................
   # scopes ....................................................................
