@@ -13,7 +13,7 @@ class ProductList < ActiveRecord::Base
   include Patcher
 
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :url, :url_key
+  # attr_accessible :url, :url_key
 
   # relationships .............................................................
   # validations ...............................................................
