@@ -2,7 +2,7 @@ class Site < ActiveRecord::Base
   # extends ...................................................................
   # includes ..................................................................
   # security (i.e. attr_accessible) ...........................................
-  attr_accessible :name
+  # attr_accessible :name
 
   # relationships .............................................................
   has_many :products
